first cleanup pass
