+++ src/ur_interface/ur_tools/grippers/abstract_gripper_interface.py
@@ -1,0 +1,20 @@
+"""Abstract gripper interface definitions."""
+
+class Gripper:
+    """Abstract gripper interface."""
+
+    def connect(self):
+        """Connect to the gripper."""
+        raise NotImplementedError("This method should be overridden by subclasses.")
+
+    def disconnect(self):
+        """Disconnect from the gripper."""
+        raise NotImplementedError("This method should be overridden by subclasses.")
+
+    def open(self, pose: float, speed: float, force: float):
+        """Open the gripper to a specified pose with given speed and force."""
+        raise NotImplementedError("This method should be overridden by subclasses.")
+
+    def close(self, pose: float, speed: float, force: float):
+        """Close the gripper to a specified pose with given speed and force."""
+        raise NotImplementedError("This method should be overridden by subclasses.")
